requirements.txt
sanic
sanic-ext
httpx
Pillow
